wip: 将vben示例的菜单与路由归入/demo中
