+++ .env
@@ -1,8 +1,8 @@
 # port
-VITE_PORT = 3100
+VITE_PORT = 3200
 
 # spa-title
-VITE_GLOB_APP_TITLE = Vben Admin
+VITE_GLOB_APP_TITLE = OSharpAdmin
 
 # spa shortname
-VITE_GLOB_APP_SHORT_NAME = vue_vben_admin
+VITE_GLOB_APP_SHORT_NAME = osharp_admin
