+++ .vscode/launch.json
@@ -5,7 +5,7 @@
       "type": "chrome",
       "request": "launch",
       "name": "Launch Chrome",
-      "url": "http://localhost:3100",
+      "url": "http://localhost:3200",
       "webRoot": "${workspaceFolder}/src",
       "sourceMaps": true
     }
