+++ package.json
@@ -9,6 +9,7 @@
   "scripts": {
     "commit": "czg",
     "bootstrap": "pnpm install",
+    "start": "npm run dev",
     "serve": "npm run dev",
     "dev": "vite",
     "build": "cross-env NODE_ENV=production vite build && esno ./build/script/postBuild.ts",
@@ -44,15 +45,19 @@
     "@vueuse/shared": "^8.3.0",
     "@zxcvbn-ts/core": "^2.0.1",
     "ant-design-vue": "^3.2.0",
+    "array-to-tree": "^3.3.2",
     "axios": "^0.26.1",
     "codemirror": "^5.65.3",
     "cropperjs": "^1.5.12",
     "crypto-js": "^4.1.1",
     "dayjs": "^1.11.1",
     "echarts": "^5.3.2",
+    "guid-typescript": "^1.0.9",
     "intro.js": "^5.1.0",
+    "linqts": "^1.14.4",
     "lodash-es": "^4.17.21",
     "mockjs": "^1.1.0",
+    "moment": "^2.29.1",
     "nprogress": "^0.2.0",
     "path-to-regexp": "^6.2.0",
     "pinia": "2.0.12",
