+++ src/locales/lang/en/routes/demo.ts
@@ -1,4 +1,5 @@
 export default {
+  demo: 'Vben Demo',
   charts: {
     baiduMap: 'Baidu map',
     aMap: 'A map',
